+++ index.js
@@ -6,7 +6,9 @@ const parserOpts = require('./src/parser-opts')
 const recommendedBumpOpts = require('./src/conventional-recommended-bump')
 const writerOpts = require('./src/writer-opts')
 
-module.exports = Q.all([conventionalChangelog, parserOpts, recommendedBumpOpts, writerOpts])
+module.exports = Q.all([ conventionalChangelog, parserOpts, recommendedBumpOpts, writerOpts ])
   .spread((conventionalChangelog, parserOpts, recommendedBumpOpts, writerOpts) => {
-    return { conventionalChangelog, parserOpts, recommendedBumpOpts, writerOpts }
+    return {
+      conventionalChangelog, parserOpts, recommendedBumpOpts, writerOpts,
+    }
   })