+++ src/parser-opts.js
@@ -1,13 +1,11 @@
 'use strict'
 
 module.exports = {
-  headerPattern: /^(\w*)(?:\((.*)\))?: (.*)$/,
-  headerCorrespondence: [
-    'type',
-    'scope',
-    'subject'
-  ],
-  noteKeywords: ['BREAKING CHANGE'],
-  revertPattern: /^(?:Revert|revert:)\s"?([\s\S]+?)"?\s*This reverts commit (\w*)\./i,
-  revertCorrespondence: ['header', 'hash']
+  headerPattern:        /^(\w*)(?:\((.*)\))?: (.*)$/,
+  // adaption for emoji-commits
+  // headerPattern:        /^(?::\w*:)\s(?<type>\w*)(?:\((?<scope>.*)\))?!?:\s(?<subject>(?:(?!#).)*(?:(?!\s).))\s?(?<ticket>#\d*)?$/,
+  headerCorrespondence: [ 'type', 'scope', 'subject', 'ticket' ],
+  noteKeywords:         [ 'BREAKING CHANGE' ],
+  revertPattern:        /^(?:Revert|revert:)\s"?([\s\S]+?)"?\s*This reverts commit (\w*)\./i,
+  revertCorrespondence: [ 'header', 'hash' ],
 }