+++ src/writer-opts.js
@@ -5,85 +5,66 @@ const Q = require('q')
 const readFile = Q.denodeify(require('fs').readFile)
 const resolve = require('path').resolve
 
-module.exports = Q.all([
-  readFile(resolve(__dirname, './templates/template.hbs'), 'utf-8'),
-  readFile(resolve(__dirname, './templates/header.hbs'), 'utf-8'),
-  readFile(resolve(__dirname, './templates/commit.hbs'), 'utf-8'),
-  readFile(resolve(__dirname, './templates/footer.hbs'), 'utf-8')
-])
-  .spread((template, header, commit, footer) => {
-    const writerOpts = getWriterOpts()
+// TODO: export for release-config (commit-analizer)
+// and try to combine with CZ
+// https://github.com/commitizen/conventional-commit-types/blob/master/index.json
+// https://github.com/commitizen/cz-conventional-changelog
 
-    writerOpts.mainTemplate = template
-    writerOpts.headerPartial = header
-    writerOpts.commitPartial = commit
-    writerOpts.footerPartial = footer
+// the actual order in the changelog...
+const types = [
+  // new features
+  { type: 'feat', section: ':sparkles: Features', release: 'minor' },
+  // bug fixes
+  { type: 'fix', section: ':beetle: Bug Fixes', release: 'patch' },
+  // configuration files - env setup (vsCode, eslint, webpack, release, travis, ...)
+  { type: 'config', section: ':wrench: Configs', release: 'patch' },
+  // user scripts (npm-scripts, husky, bash, gulp/grunt,...)
+  { type: 'script', section: ':hammer_and_wrench: Tasks and Scripts', release: 'patch' },
+  // themes, UI and assets
+  { type: 'style', section: ':art: Styling', release: 'patch' },
+  // dependencies - NPM packages and other 3rd party resources
+  { type: 'deps', section: ':package: Dependencies', release: 'patch' },
 
-    return writerOpts
-  })
+  // NO VERSION BUMP:
+  //
+  // refactoring - structure improvements
+  { type: 'refactor', section: ':building_construction: Refactoring', release: false },
+  // karma, mocha/chai/enzyme, cypress - setup and specations
+  { type: 'test', section: ':white_check_mark: Test System', release: false },
+  // changes in documentations - README, WIKI, CHANGELOG, ...
+  { type: 'docs', section: ':memo: Documentations', release: false },
+  // changes in dependencies - add, upgrade, remove, ...
 
-function getWriterOpts () {
+  // revert status - TODO: how to use
+  { type: 'revert', section: ':rewind: Reverts', release: false },
+]
+// achieved throug orderMap
+const sectionOrder = new Map(types.map((o, i) => o.section).map((s, i) => [ s, i ]))
+// and custom sorting by map keys
+function commitGroupsSort(groupA, groupB) {
+  const rankA = sectionOrder.has(groupA.title) ? sectionOrder.get(groupA.title) : 1000
+  const rankB = sectionOrder.has(groupB.title) ? sectionOrder.get(groupB.title) : 1000
+  return rankA - rankB
+}
+// easy acces to titles
+const sectionsTitles = {} // = _.indexBy(types, 'type')
+types.map((o, i) => { sectionsTitles[o.type] = o.section })
+
+function getWriterOpts() {
   return {
+    types,
     transform: (commit, context) => {
-      let discard = true
+      // let discard = true
       const issues = []
 
-      commit.notes.forEach(note => {
+      commit.notes.forEach((note) => {
         note.title = 'BREAKING CHANGES'
-        discard = false
+        // discard = false
       })
-      
-      // set section title
-      const types = [
-        { type: "feat", section: ":sparkles: Features"},
-        { type: "fix", section: ":bug: Bug Fixes" },
-        { type: "refactor", section: ":building_construction: Refactoring"},
-        { type: "style", section: ":art: Styling" },
-        { type: "docs", section: ":memo: Documentations" },
-        { type: "perf", section: ":zap: Performance Enhancement" },
-        { type: "revert", section: ":rewind: Reverts" },
-        { type: "test", section: ":white_check_mark: Tests" },
-        { type: "build", section: ":construction_worker: Continuous Integration" },
-        { type: "ci", section: ":hammer_and_wrench: Build System" },
-      ]
-      
-      const sectionsTitles = {}
-      types.map((o, i) => { sectionsTitles[o.type] = o.section })
-      
-      console.log(sectionsTitles)
-      
-      let sectionOrder = new Map(
-        types.map((o, i) => o.section)
-          .map((s, i) => [s, i])
-      )
-      
-      console.log(sectionOrder)
-      
-      function commitGroupsSort(groupA, groupB) {
-        const rankA = sectionsOrder.has(groupA.title) ? sectionsOrder.get(groupA.title) : 1000;
-        const rankB = sectionsOrder.has(groupB.title) ? sectionsOrder.get(groupB.title) : 1000;
-      
-        return rankA - rankB;
-      }
-      
-      function compareStrings(a, b) {
-        if (a < b) return -1;
-        if (a > b) return 1;
-      
-        return 0;
-      }
-      
-      function commitsSort(commitA, commitB) {
-        const scopeRank = compareStrings(commitA.scope, commitB.scope);
-      
-        if (scopeRank === 0) return compareStrings(commitA.subject, commitB.subject);
-        if (!commitA.scope) return 1;
-        if (!commitB.scope) return -1;
-      
-        return scopeRank;
-      }
 
-      commit.type = sectionsTitles[commit.type]
+      // set type group titles
+      if (sectionsTitles[commit.type] == null) return
+      commit.type = sectionsTitles[commit.type] || commit.type
 
       if (commit.scope === '*') {
         commit.scope = ''
@@ -118,7 +99,7 @@ function getWriterOpts () {
       }
 
       // remove references that already appear in the subject
-      commit.references = commit.references.filter(reference => {
+      commit.references = commit.references.filter((reference) => {
         if (issues.indexOf(reference.issue) === -1) {
           return true
         }
@@ -128,10 +109,27 @@ function getWriterOpts () {
 
       return commit
     },
-    groupBy: 'type',
-    commitGroupsSort, // 'title'
-    commitsSort: ['scope', 'subject'],
+    groupBy:        'type',
+    commitGroupsSort,
+    commitsSort:    [ 'scope', 'subject' ],
     noteGroupsSort: 'title',
-    notesSort: compareFunc
+    notesSort:      compareFunc,
   }
 }
+
+module.exports = Q.all([
+  readFile(resolve(__dirname, '../templates/template.hbs'), 'utf-8'),
+  readFile(resolve(__dirname, '../templates/header.hbs'), 'utf-8'),
+  readFile(resolve(__dirname, '../templates/commit.hbs'), 'utf-8'),
+  readFile(resolve(__dirname, '../templates/footer.hbs'), 'utf-8'),
+])
+  .spread((template, header, commit, footer) => {
+    const writerOpts = getWriterOpts()
+
+    writerOpts.mainTemplate = template
+    writerOpts.headerPartial = header
+    writerOpts.commitPartial = commit
+    writerOpts.footerPartial = footer
+
+    return writerOpts
+  })