+++ CHANGELOG.md
@@ -1,7 +1,88 @@
-## <small>5.0.12 (2021-05-14)</small>
+# Changelog
 
-* feat: add emojis to type section headlines ([2492d11](https://github.com/conventional-changelog/conventional-changelog/commit/2492d11))
-* inital commit ([c71ce4a](https://github.com/conventional-changelog/conventional-changelog/commit/c71ce4a))
+> generated with [@semantic-release/changelog](https://github.com/semantic-release/changelog)
 
+## [1.1.6](https://github.com/stackr23/preset-conventional-changelog/compare/v1.1.5...v1.1.6) (2021-05-17)
 
+### :beetle: Bug Fixes
 
+- change type definitions ([f8242e7](https://github.com/stackr23/preset-conventional-changelog/commit/f8242e7273458b8e5c6286d66b34e057b0b30841))
+- **release:** remove 'conventionalcommits' preset ([2f94cd6](https://github.com/stackr23/preset-conventional-changelog/commit/2f94cd65d2a09dca91a78e73fc8a30590fb765e7))
+
+### [1.1.5](https://github.com/stackr23/preset-conventional-changelog/compare/v1.1.4...v1.1.5) (2021-05-16)
+
+### Bug Fixes
+
+- **vsCode:** dont format handlebars files ([3e6735b](https://github.com/stackr23/preset-conventional-changelog/commit/3e6735b24338d7632ddb779968991620c3ac4572))
+- reorder type groups ([70c8c3a](https://github.com/stackr23/preset-conventional-changelog/commit/70c8c3a223470240270ba3b7936e691afc840822))
+- **templates:** use relative paths ([3a44c6e](https://github.com/stackr23/preset-conventional-changelog/commit/3a44c6e201f9199544153675836c8cfb97296e2d))
+
+## [1.1.4](https://github.com/stackr23/preset-conventional-changelog/compare/v1.1.3...v1.1.4) (2021-05-15)
+
+### :beetle: Bug Fixes
+
+- **npm:** use right pattern for src files ([1819f22](https://github.com/stackr23/preset-conventional-changelog/commit/1819f2292197b1bf56980a0b700e89ea8c700443))
+
+### :memo: Documentations
+
+- **CHANGELOG:** add missing emoji in headline ([b1fb4c3](https://github.com/stackr23/preset-conventional-changelog/commit/b1fb4c348d1d72d49a12ff520247c1e01654b3d3))
+
+## [1.1.3](https://github.com/stackr23/preset-conventional-changelog/compare/v1.1.2...v1.1.3) (2021-05-15)
+
+### :beetle: Bug Fixes
+
+- **npm:** add src files to pckg ([7c3c7b8](https://github.com/stackr23/preset-conventional-changelog/commit/7c3c7b889723a7003d6c4e972f9c57d5401996c9))
+
+### :wrench: Configs
+
+- **release:** dont patch on 'docs' ([3e33aa6](https://github.com/stackr23/preset-conventional-changelog/commit/3e33aa6df8d79b0a6b193615dfb065009f733a86))
+
+## [1.1.2](https://github.com/stackr23/preset-conventional-changelog/compare/v1.1.1...v1.1.2) (2021-05-14)
+
+### :wrench: Configs
+
+- **travis:** install '@stackr23/preset-conventional-changelog' globally ([ab67bce](https://github.com/stackr23/preset-conventional-changelog/commit/ab67bcee2d4a67cf8eedb7869a319e613d285e9e))
+
+### :memo: Documentations
+
+- **README:** add type adaptions ([0e7b4b4](https://github.com/stackr23/preset-conventional-changelog/commit/0e7b4b44e08750b7ef3ab127a4106e9f8a6f050f))
+
+## [1.1.1](https://github.com/stackr23/preset-conventional-changelog/compare/v1.1.0...v1.1.1) (2021-05-14)
+
+### :beetle: Bug Fixes
+
+- **types:** change 'ci' and 'build' titles ([ef07a87](https://github.com/stackr23/preset-conventional-changelog/commit/ef07a8764ca538607747020c94b2ff21dca16cf9))
+
+### :building_construction: Refactoring
+
+- change type group order ([407ee53](https://github.com/stackr23/preset-conventional-changelog/commit/407ee532557d160de8c517c331bd4c2e755462be))
+- move module code to `/src` ([cfc4741](https://github.com/stackr23/preset-conventional-changelog/commit/cfc4741d9e0cc1829432e5e9e6c60c4ec6bfb6af))
+
+### :wrench: Configs
+
+- **travis:** add default travis config ([d5b73f4](https://github.com/stackr23/preset-conventional-changelog/commit/d5b73f4694212a31a3282cdc986639aa5e2054b0))
+- **travis:** install '@stackr23/preset-conventional-changelog' before deployment ([1ead94a](https://github.com/stackr23/preset-conventional-changelog/commit/1ead94a0dfc1ee0f2682197a556735773bea6575))
+- **release:** add 'semantic-release' config ([932e342](https://github.com/stackr23/preset-conventional-changelog/commit/932e3427deb49f53689ddff47377e7dcb8075344))
+
+### :memo: Documentations
+
+- **README:** add npm package badge ([da98693](https://github.com/stackr23/preset-conventional-changelog/commit/da9869338d5d35eea530b3d3ff58771e31754caa))
+
+# [1.1.0](https://github.com/stackr23/preset-conventional-changelog/compare/7cf35aa8315ec9713f38a75110af62b85e86ffc1...v1.1.0) (2021-05-14)
+
+### :sparkles: Features
+
+- add eslint and prettier ([a7782ea](https://github.com/stackr23/preset-conventional-changelog/commit/a7782ea40c7ceae5f29d3e33d425ffc822f3dace))
+
+### :beetle: Bug Fixes
+
+- **sortOrder:** fix `sectionOrder` Map ([95a2030](https://github.com/stackr23/preset-conventional-changelog/commit/95a2030a5fe6a7a5693da4536e53166199d52292))
+- sort commit gropus by types array ([ece59d5](https://github.com/stackr23/preset-conventional-changelog/commit/ece59d5fbe73701d39d5454e5c9b6dd01f5ae51c))
+
+### :building_construction: Refactoring
+
+- define and apply section titles with emojis ([7cf35aa](https://github.com/stackr23/preset-conventional-changelog/commit/7cf35aa8315ec9713f38a75110af62b85e86ffc1))
+
+### :memo: Documentations
+
+- add CHANGELOG ([ed5e1f7](https://github.com/stackr23/preset-conventional-changelog/commit/ed5e1f7d583455bfa597f77be70dec18cfef2a59))